+++ php/src-generated/TestStepResult.php
@@ -26,8 +26,17 @@ final class TestStepResult implements JsonSerializable
      */
     public function __construct(
         public readonly Duration $duration = new Duration(),
+
+        /**
+         * An arbitrary bit of information that explains this result. This can be a stack trace of anything else.
+         */
         public readonly ?string $message = null,
         public readonly TestStepResult\Status $status = TestStepResult\Status::UNKNOWN,
+
+        /**
+         * Exception thrown while executing this step, if any.
+         */
+        public readonly ?Exception $exception = null,
     ) {
     }
 
@@ -41,11 +50,13 @@ final class TestStepResult implements JsonSerializable
         self::ensureDuration($arr);
         self::ensureMessage($arr);
         self::ensureStatus($arr);
+        self::ensureException($arr);
 
         return new self(
             Duration::fromArray($arr['duration']),
             isset($arr['message']) ? (string) $arr['message'] : null,
             TestStepResult\Status::from((string) $arr['status']),
+            isset($arr['exception']) ? Exception::fromArray($arr['exception']) : null,
         );
     }
 
@@ -82,6 +93,16 @@ final class TestStepResult implements JsonSerializable
         }
         if (array_key_exists('status', $arr) && is_array($arr['status'])) {
             throw new SchemaViolationException('Property \'status\' was array');
+        }
+    }
+
+    /**
+     * @psalm-assert array{exception?: array} $arr
+     */
+    private static function ensureException(array $arr): void
+    {
+        if (array_key_exists('exception', $arr) && !is_array($arr['exception'])) {
+            throw new SchemaViolationException('Property \'exception\' was not array');
         }
     }
 }